app/requirements.txt
streamlit==1.48.0
torch==2.8.0
torchvision==0.23.0
Pillow==11.3.0
PyYAML==6.0.2
scikit-learn==1.7.1
matplotlib==3.10.5
seaborn==0.13.2
tqdm==4.67.1
